+++ website/src/Components/LayoutContainer/LayoutContainer.tsx
@@ -51,14 +51,14 @@ const LayoutContainer = ({ children, ...props }: LayoutProps) => {
       {...props}
     >
       <HeadSeo title={currentComponent?.name} description="Hooks utilities." />
-      <AsideStyle className="hidden lg:block fixed z-20 inset-0 top-[88px] left-[max(0px,calc(50%-45rem))] right-auto w-[19.5rem] py-10 px-8 overflow-y-auto">
+      <AsideStyle className="hidden lg:block fixed z-20 inset-0 top-[88px] left-[max(0px,calc(50%-56.5rem))] right-auto w-[19.5rem] py-10 px-8 overflow-y-auto">
         <nav className="lg:text-sm lg:leading-6 relative h-full px-5">
           <Aside t={t} />
         </nav>
       </AsideStyle>
 
-      <BodyStyle className="lg:pl-[19.5rem]">
-        <ContentStyle className="max-w-3xl mx-auto pt-10 xl:max-w-none xl:ml-0 xl:mr-[13.2rem] xl:pr-16">
+      <BodyStyle className="lg:pl-[17rem]">
+        <ContentStyle className="max-w-3xl mx-auto pt-10 xl:max-w-none xl:ml-7 xl:mr-[13.2rem] xl:pr-4">
           {children}
           <div className="h-20" />
           <div className="flex justify-between items-end">
@@ -82,7 +82,7 @@ const LayoutContainer = ({ children, ...props }: LayoutProps) => {
           </div>
           <Footer className="rounded p-4" completeFooter={false} />
 
-          <BreadStyle className="fixed z-20 top-[3.8125rem] bottom-0 right-[max(0px,calc(50%-45rem))] w-[19.5rem] py-10 overflow-y-auto hidden xl:block">
+          <BreadStyle className="fixed z-20 top-[3.8125rem] bottom-0 right-[max(0px,calc(50%-58rem))] w-[19.5rem] py-10 overflow-y-auto hidden xl:block">
             <div className="px-8">
               <TableOfContent headings={headings} />
             </div>